requirements.txt
anyio==3.6.1
certifi==2022.9.14
cffi==1.15.0
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
fastapi==0.85.0
fastapi-cache2==0.1.9
gTTS==2.2.4
hpy==0.0.3
idna==3.4
pydantic==1.10.2
requests==2.28.1
six==1.16.0
sniffio==1.3.0
starlette==0.20.4
typing_extensions==4.3.0
urllib3==1.26.12